Tema4/requirements.txt
click==6.7
Flask==1.1.1
itsdangerous==0.24
Jinja2==2.10.1
MarkupSafe==1.0
Werkzeug==0.15.5
PyMongo==3.10.1
pycryptodomex==3.9.7
azure-storage-blob==12.3.0
azure-cognitiveservices-vision-computervision==0.5.0